Some refactorings.
