+++ DB/src/main/java/io/deephaven/db/v2/locations/parquet/local/ParquetColumnLocation.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2021 Deephaven Data Labs and Patent Pending
  */
 
-package io.deephaven.db.v2.locations.local;
+package io.deephaven.db.v2.locations.parquet.local;
 
 import io.deephaven.base.verify.Require;
 import io.deephaven.configuration.Configuration;
